Implemented a barebones httpServer by deriving from TcpServer class. The server works well. Updated README.
messageHandler is implemented in the derived class.
The messageHandler receives messages from concurrent message que, which is member of the base class, by using receiveMessage() method of the base class.
Since messages are stored in the message que as a pair of clientSocket and message from client, the message handler can call Send() method of the base calls to respond to the messages.